Ensure SequelRails::Configuration is used everywhere to have correct default (fixes #44)
Previously, `app.config.sequel` was first assigned an
`ActiveSupport::OrderedOptions` on `Railtie` loading, and then during
the Rails initialization phase it was replaced by a
`SequelRails::Configuration` object (this is something which have bypass
my vigilance). When loading the rake's tasks, the initializers are not
run, only the Railtie's `rake_tasks` hook is, which means the defaults
were not applied correctly, resulting in the SequelRails database task
to not be loaded automatically.